Publish SDK artifacts as aar not jar
The Android ARchieve (aar) format has been around for a long time. The
Android SDK originated from a fork of the AWS Java SDK, which was
publishing Java ARchives (jar). While it possible to conusme jar files
on Android, it is not idiomatic to the platform. Instead, we would like
to publish Android libraries.

This change may require consumers to update their build.gradle, in such
cases as they have '@jar' notation in their dependency specifications.