[aws-android-sdk-testutils] Remove testconfiguration.json. (#1567)
This file is listed in .gitignore, but a templated version exists in the
committed source code. In order for changes to this file to _actually_
be ignored from versioning, this file must not be tracked by git.

This means that R.raw.testconfiguration will no longer be available at
compile time. To address this, the AWSTestBase is modified to lookup the
configuration file by its name, at runtime.