Stop using key pair in autoscaling tests (#1637)
Previously, these tests used a hard-coded keypair that was expected to
be in the test account named by the alias of the developed that worked
on this feature. EC2 key pairs cannot be created via CloudFormation, but
aren't required in launch configurations. There's minimal impact to
these tests by removing this element and assertion.