Ignore topic modeling integration test
The Comprehend suite has an elaborate test that runs topics detection
which allows Comprehend to example a collection of documents to
determine common themes. There isn't a clear mobile use case for this as
topic modeling jobs are recommended to be ran with at least 1,000
documents and are run asynchronously. This is more of a batch operation,
and the test has a 20 minute maximum runtime to reflect that. Given the
dubious nature of the mobile use case, the resources required to
orchestrate the test (IAM Role, S3 Bucket), and the amount of time it
takes to run, I'm putting @Ignore on it versus snaking through the AWS
resources to rehabilitate it.