Remove allowBackup="true" from AndroidManifest.xml
The AndroidManifest.xml in aws-android-sdk-cognitoauth has `allowBackup=true`. That decision should not belong to a dependency, and having this entry creates friction for consumers of the SDK as it can generate build errors:

```
Manifest merger failed : Attribute application@allowBackup value=(false) from [:common:mymodule] AndroidManifest.xml:37:9-36
        is also present at [com.amazonaws:aws-android-sdk-cognitoauth:2.16.12] AndroidManifest.xml:4:19-45 value=(true).
        Suggestion: add 'tools:replace="android:allowBackup"' to <application> element at AndroidManifest.xml:4:3-5:41 to override.
```

While consumers can use `tools:replace`, it's best to just not have this and avoid the issue in the first place.