chore: split iot tests into their own module (#1940)
The usual pattern in this repository is to pair an SDK module with a
separate test module. The `aws-android-sdk-iot` module did not fit this
pattern; source, unit, and insturmentation tests live in the same
module.

This commit splits out the on-device instrumentation tests from the
`aws-android-sdk-iot` module, and into a new `aws-android-sdk-iot-test`
module.