chore: use latest dependency versions (#1997)
The SDK has not been receiving regular updates to its dependencies, and
so many of them have drifted out of date. This large "catch up" commit
updates all dependencies to the latest versions in Maven.

The `build.gradle` and `pom.xml` is updated to include the latest
versions available through Maven.

This change will effect the consumer's dependency graph, and may be
considered a breaking change. It will go into the 2.17.0 release.