fix: derrive min sdk values from android lint
For the purpose of re-enabling NewApi and other API-related Android Lint
warnings, all abortOnError options are removed. In response, some
unrelated lint warnings are being addressed or suppressed in code, as a
result of this commit.

To derive an appropriate minSdkVersion for each module, minSdkVersion
was stripped from all modules, and then updated according to the project
build requirements.