fix: remove apache http use from s3 (#2009)
The S3 SDK contained deprecated code which stored an Apache HTTP object
in the S3ObjectInputStream. The time has come to make good of the
deprecation notice. Several dead constructors and a getter are being
removed, so that the dependency on Apache HTTP may fully be removed from
the S3 module.