chore: remove jackson support from sdk core
The code generated components of the SDK are written in such a way as to
support different serialization engines. However, the hand-written
clients that sit atop them include hardcoded usage of Gson. Amplify
itself makes opinionated use of Gson. It is not clear that there is any
value to support Jackson, any more. Certainly, it should not be a
required dependency of core, whilst Gson is also required. To resolve
this, Jackson support is removed and Gson is set as the only
serialization runtime.