chore: remove legacy apache http from core (#2014)
The Android SDK uses HttpURLConnection as its HTTP runtime.
aws-android-sdk-core contains dead code that references the legacy
Apache HTTP Client used by the V1 Java SDK, from which this codebase
originates. The Apache HTTP client has considerable baggage on the
Android platform, and can cause runtime failures. In an effort to fully
remove it from the codebase, it is being removed from core.

As of this commit, Apache Logging is still a compile-time dependency in
core. If it is provided at runtime, it will be used. Otherwise, the SDK
will fallback to logging against the Android Log utility. This behavior
is discussed here: https://github.com/aws-amplify/aws-sdk-android/issues/476