fix: remove unnecessary contextcompat class lookup (#2091)
AWSMobileClient tried to check if the library has network access. As
part of this, it checks whether or not there is permission to access the
network. The existing mechanism to achieve this performed a reflective
class lookup on an (old) Android support library.

Instead, guard the checkSelfPermission(..) call in an API level guard.

Resolves: https://github.com/aws-amplify/aws-sdk-android/issues/2063