fix(aws-android-sdk-iot): Improve AWS Region code extraction (#2325)
* Improve AWS Region code extraction

Previously this helper was looking for an AWS Region code within a
static component of the endpoint. AWS IoT Device Advisor introduces a
new format for test endpoints. Rather than adding to this logic, this
change tokenizes the endpoint and looks for a matching AWS Region code.

- Improve and tidy up the tests a bit

* Eliminate unneeded helper class

This helper class is only used in a single place, so instead of
maintaining a standalone class with one static method move it into the
class using it.